New views file for spotifyAPP which handles the GET and Redirect
This file contains does two functions at the moment. It calls a get to request the url for spotify to receive tokens for the user, this will lead the user to sign in to authenticate these tokens. Then Once the tokens have been gotten and are authenticated, we store the tokens. We store the tokens i  the Model.py file for spotify and stored in our db. Then once we have stored these tokens and the user is authenticated, we can redirect our user back to their app. The data is stored, and the user is authenticated.